refactor v2 trace keyed collection ops
